fix(web): resolve ESLint errors in TraceViewer component
Move buildTree utility to separate lib/trace-utils.ts to fix
react-refresh/only-export-components rule. Replace dynamic icon
component reference with static ActionIcon component to fix
react-hooks/static-components rule.
